requirements.txt
apache-airflow
requests
beautifulsoup4
psycopg2-binary
